GenAI/rag/requirements.txt
faiss-cpu
numpy
requests
langchain-ollama
